new: new decorator implementation
This implementation is way smaller, and solve technical problems of
the previous implementation. The most important limitation is that
because we'd trigger the actual decorator upon first execution of the
function, we could not prevent the rest of the function to run.

The cost of removing this limitation is that we need to explicitly
call for the mangling of the function being decorated. This isn't
yet automated but could be part of ``include`` at some point.

Signed-off-by: Valentin Lab <valentin.lab@kalysto.org>